+++ src/main/java/com/j9/bestmoments/service/oauthService/OAuthService.java
@@ -1,4 +1,4 @@
-package com.j9.bestmoments.service;
+package com.j9.bestmoments.service.oauthService;
 
 import com.j9.bestmoments.dto.response.OAuthUserInfoDto;
